Revert "#288 Apresentar Repositorios de Colaborador"
