+++ .gitignore
@@ -67,6 +67,4 @@ venv.bak/
 export.sh
 
 export.sh
-loaded-env.txt
-outputs/*
 # End of https://www.gitignore.io/api/macos,python,visualstudiocode
