+++ api/gitlab/__init__.py
@@ -6,7 +6,6 @@ from gitlab.build.views import build_blueprint
 from gitlab.webhook.views import webhook_blueprint
 from gitlab.user.views import user_blueprint
 from gitlab.rerun_pipeline.views import rerun_pipeline_blueprint
-from gitlab.stable_deploy.views import stable_deploy_blueprint
 from flask_cors import CORS
 
 cors = CORS()
@@ -30,7 +29,6 @@ def create_app(script_info=None):
     app.register_blueprint(webhook_blueprint)
     app.register_blueprint(user_blueprint)
     app.register_blueprint(rerun_pipeline_blueprint)
-    app.register_blueprint(stable_deploy_blueprint)
 
     # shell context for flask cli
     @app.shell_context_processor
