+++ api/gitlab/cronjob/crontab
@@ -1,0 +1,1 @@
+* * * * *  ./api/gitlab/cronjob/scripts/monitoring_domain.py >> ~/cron.log 2>&1
