--- api/gitlab/stable_deploy/error_messages.py
@@ -1,2 +1,0 @@
-NOT_FOUND = {"message": "NOT FOUND", "status_code": 404}
-UNAUTHORIZED = {"message": "UNAUTHORIZED", "status_code": 401}
