--- api/gitlab/tests/jsonschemas/stable_deploy/schemas.py
@@ -1,60 +1,0 @@
-ping_schema = {
-    "$schema": "http://json-schema.org/draft-04/schema#",
-    "title": "User response schema",
-    "type": "object",
-    "properties": {
-        "message": {"type": "string"},
-        "status": {"type": "string"}
-    },
-    "required": ["message", "status"]
-}
-
-valid_schema = {
-    "$schema": "http://json-schema.org/draft-04/schema#",
-    "title": "User response schema",
-    "type": "array",
-    "properties": {
-        "status_code": {"type": "integer"}
-    },
-    "required": ["status_code"]
-}
-
-unauthorized_schema = {
-    "$schema": "http://json-schema.org/draft-04/schema#",
-    "title": "User response schema",
-    "type": "object",
-    "properties": {
-        "status_code": {"type": "integer"}
-    },
-    "required": ["status_code"]
-}
-
-stable_deploy_schema = {
-    "$schema": "http://json-schema.org/draft-04/schema#",
-    "title": "User response schema",
-    "type": "object",
-    "properties": {
-        "status": {"type": "string"}
-    },
-    "required": ["status"]
-}
-
-pipeline_id_scheema = {
-    "$schema": "http://json-schema.org/draft-04/schema#",
-    "title": "User response schema",
-    "type": "object",
-    "properties": {
-        "status": {"type": "string"}
-    },
-    "required": ["status"]
-}
-
-invalid_project_schema = {
-    "$schema": "http://json-schema.org/draft-04/schema#",
-    "title": "User response schema",
-    "type": "object",
-    "properties": {
-        "status_code": {"type": "integer"}
-    },
-    "required": ["status_code"]
-}
