+++ api/gitlab/tests/test_user.py
@@ -5,8 +5,7 @@ from gitlab.tests.jsonschemas.user.schemas import\
     unauthorized_schema,\
     user_data_valid_schema, project_id_schema,\
     user_id_schema, get_user_domain_schema,\
-    save_user_domain_schema, user_invalid_schema,\
-    get_user_infos_schema
+    save_user_domain_schema, user_invalid_schema
 from jsonschema import validate
 from gitlab.user.utils import UserUtils, send_message
 from requests import Response
@@ -260,14 +259,6 @@ class TestUser(BaseTestCase):
         mocked_response.status_code = 200
         mocked_post.return_value = mocked_response
         authenticate_access_token("44456")
-
-    def test_views_get_user_infos(self):
-        chat_id = self.user.chat_id
-        response = self.client.get("/user/infos/{chat_id}"
-                                   .format(chat_id=chat_id))
-        data = json.loads(response.data.decode())
-        self.assertEqual(response.status_code, 200)
-        validate(data, get_user_infos_schema)
 
 
 if __name__ == "__main__":
