--- cronjob/entrypoint_cron.sh
@@ -1,2 +1,0 @@
-#!/bin/sh
-cron -f
