--- cronjob/update_project_model.sh
@@ -1,5 +1,0 @@
-#!/bin/sh
-tail -n +2 "project.py" > "project.tmp" && mv "project.tmp" "project.py"
-echo 'from __init__ import init_db' > temp_file.py
-cat project.py >> temp_file.py
-mv temp_file.py project.py
