+++ docker-compose.yml
@@ -31,18 +31,7 @@ services:
       - 20191-ada_adanetwork
     ports:
       - 27010:27010
-  cronjob:
-    image: joaovitor3/ada-gitlab-cronjob:homolog
-    volumes:
-      - ./outputs:/cronjob/home
-    stdin_open: true
-    tty: true
-    networks:
-      - 20191-ada_adanetwork
-    environment:
-      - DB_URL=${DB_URL}
-      - DB_NAME=${DB_NAME}
-      - ACCESS_TOKEN=${ACCESS_TOKEN}
+
 networks:
   20191-ada_adanetwork:
     external: true
