--- docker/cronjob.requirements.txt
@@ -1,3 +1,0 @@
-python-telegram-bot==11.1.0
-mongoengine==0.16.3
-requests==2.21.0
