+++ docker/dev.Dockerfile
@@ -4,10 +4,6 @@ FROM python:3.6-slim
 # set working directory
 WORKDIR /usr/src/app
 
-# install cronjob dependencies
-RUN apt-get update && apt-get -y install nano \
-    && apt-get -y install cron
-
 # add and install requirements
 ADD ./docker/requirements.txt /usr/src/app/requirements.txt
 RUN pip install -r requirements.txt
